third_party/blink/web_tests/http/tests/security/cors-check-for-cached-script.html

<!doctype html>
<html>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="http://localhost:8080/security/resources/cors-script.php?max-age=60" crossorigin></script>
<script>
async_test((test) => {
  window.addEventListener('message', test.step_func_done((e) => {
    assert_equals(e.data, 'PASS');
  }));
}, 'We should not reuse CORS check result when requested from another origin');
</script>
<iframe sandbox="allow-scripts allow-modals" src="resources/cors-check-for-cached-script-iframe.html"></iframe>
